@@ -3,9 +3,9 @@ package com.spring.boot.connection.application;
 import com.spring.boot.common.exception.NotFoundException;
 import com.spring.boot.connection.domain.ConnectionRepository;
 import com.spring.boot.connection.domain.Connection;
-import com.spring.boot.user.application.dto.UserInfoDto;
-import com.spring.boot.user.domain.Member;
-import com.spring.boot.user.domain.UserRepository;
+import com.spring.boot.member.application.dto.UserInfoDto;
+import com.spring.boot.member.domain.Member;
+import com.spring.boot.member.domain.UserRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
