@@ -1,14 +1,14 @@
 package com.spring.boot.post.domain;
 
 import com.spring.boot.common.BaseTime;
-import com.spring.boot.user.domain.Member;
+import com.spring.boot.like.domain.Like;
+import com.spring.boot.like.domain.Likes;
+import com.spring.boot.member.domain.Member;
 import com.spring.boot.post.domain.image.PostImage;
 import com.spring.boot.post.domain.image.PostImages;
-import com.spring.boot.post.domain.like.PostLike;
 import com.spring.boot.post.domain.tag.PostTag;
 import com.spring.boot.post.domain.tag.PostTags;
 import com.spring.boot.tag.domain.Tag;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import javax.persistence.Column;
@@ -20,13 +20,11 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
-import org.hibernate.annotations.BatchSize;
 
 @Entity
 @Getter
@@ -54,9 +52,10 @@ public class Post extends BaseTime {
   @Embedded
   private PostTags postTags = new PostTags();
 
-  @BatchSize(size = 1000)
-  @OneToMany(mappedBy = "post")
-  private List<PostLike> postLikes = new ArrayList<>();
+  @Embedded
+  private Likes likes = new Likes();
+
+  private int likeTotalCount;
 
   public Post(String title, String content, Member writer) {
     this.title = title;
@@ -82,12 +81,27 @@ public class Post extends BaseTime {
     initPostTags(tags);
   }
 
+  public void like() {
+    this.likeTotalCount += 1;
+  }
+
+  public void unlike(){
+    if(this.likeTotalCount - 1 < 0){
+      throw new IllegalStateException();
+    }
+    this.likeTotalCount -= 1;
+  }
+
   public Set<PostTag> getPostTags() {
     return postTags.getPostTags();
   }
 
   public List<PostImage> getPostImages() {
     return postImages.getPostImages();
+  }
+
+  public List<Like> getLikes() {
+    return likes.getLikes();
   }
 
   @Override
