@@ -4,8 +4,8 @@ import com.spring.boot.common.DatabaseCleanUp;
 import com.spring.boot.common.config.InfrastructureTestConfiguration;
 import com.spring.boot.common.mock.MockMember;
 import com.spring.boot.common.mock.MockPost;
-import com.spring.boot.user.domain.Member;
-import com.spring.boot.user.domain.UserRepository;
+import com.spring.boot.member.domain.Member;
+import com.spring.boot.member.domain.UserRepository;
 import com.spring.boot.post.domain.Post;
 import com.spring.boot.post.infrastructure.PostRepository;
 import com.spring.boot.role.domain.Role;
@@ -40,14 +40,17 @@ public abstract class IntegrationTest {
   protected Member saveMember(String email) {
     Role role = roleRepository.findByRoleName(RoleName.MEMBER.getValue())
         .orElseGet(() -> roleRepository.save(new Role(RoleName.MEMBER.getValue(), "member")));
-    Member member = MockMember.builder(email, "name", "password", role)
+
+    Member member = MockMember.builder(email, role)
         .build();
+
     return userRepository.save(member);
   }
 
   protected Post savePost(Member writer) {
     Post post = MockPost.builder("title", "content", writer)
         .build();
+
     return postRepository.save(post);
   }
 
